Move json related classes into a Json namespace
